ROCm-OpenCL-Runtime: specify source directory names
Previously, ROCm-OpenCL-Runtime relied on fetchFromGitHub using an
escaped version of ${repo}-${rev}-src as the name of the downloaded
directory by default. In nixpkgs master, fetchFromGitHub no longer
does this, and instead defaults to calling the downloaded directories
"source", which causes issues when the unpacker attempts to unpack
multiple at once. This commit. specifies the name for each downloaded
repository, fixing this issue.