Add hcc, hip, and clang-ocl; refactor rocm-opencl-runtime
Break apart the rocm-opencl-runtime build to separate the builds of
LLVM and Clang. When something goes wrong, this makes the resulting
re-build somewhat less painful.

The family of ROCm repositories includes multiple forks of
clang (`clang` and `hcc-clang-upgrade`), but
isolating the builds of these tools lets us better take advantage of
re-use between the various ROCm compilers (i.e. OpenCL, hcc, and hip).

The build of hcc (AMD's single-source C++ compiler) is made more
complicated by virtue of the fact that it uses the clang it builds to
build some of its own tools. However, for that compiler to find system
includes and libraries, we need to use the nixpkgs cc-wrapper
tooling. That wrapping must happen part way through the upstream hcc
build (after the build of clang, but before more AMD-specific
libraries for things like AMP); splitting out the builds of LLVM and
clang makes it possible to interpose this wrapping.

The small clang-ocl project is useful for offline compilation
of OpenCL kernels into object files that can be loaded by downstream
consumers.

Finally, this should reflect a better state of the staggered roc-1.9.0
release across the many ROCm repositories, not all of which were ready
at the time of initial release.