rocm-opencl-runtime: address upstream build bug
Jakov Kholodkov noted that rocm-opencl-runtime *requires* parallel
building. This was initially satisfied by having
`enableParallelBuilding = true;` in the .nix file, but Jakov pointed out that
even with that setting, one must ensure that nix uses multiple
builders (e.g. by setting `nix.buildCores = 4;` in
`/etc/nixos/configuration.nix`).

After looking into it more, it seems that the `oclruntime` target
needs an explicit dependency on the `oclrocm` target. With this
dependency in place, I was able to build with `enableParallelBuilding
= false;` (or `make -j 1` in a `nix-shell '<nixpkgs>' -A
rocm-opencl-runtime`).