hip: move patchPhase commands to postPatch
The hip derivation used patchPhase for fixing up paths. However, this
overrides the default patchPhase of stdenv. This makes it harder to
override the hip derivation with a `patches` attribute, since patches
won't be applied by patchPhase. This change moves the path fixups to
postPatch, so that patchPhase of stdenv can still be used.